fixed recording, need testing
