+++ modules/bag/CMakeLists.txt
@@ -16,8 +16,6 @@ set(SOURCES
     src/zenoh_recorder.cpp
     src/topic_filter.cpp
     src/writer.cpp
-    src/zenoh_recorder_subscribers_manager.h
-    src/zenoh_recorder_subscribers_manager.cpp
     README.md
     include/eolo/bag/topic_filter.h
     include/eolo/bag/zenoh_recorder.h
