+++ modules/bag/apps/bag_recorder.cpp
@@ -1,10 +1,8 @@
-//=================================================================================================
 // Copyright (C) 2023-2024 EOLO Contributors
 //=================================================================================================
 
 #include <csignal>
 
-#include "eolo/bag/writer.h"
 #include "eolo/bag/zenoh_recorder.h"
 #include "eolo/cli/program_options.h"
 #include "eolo/ipc/common.h"
@@ -23,23 +21,20 @@ auto main(int argc, const char* argv[]) -> int {
   desc.defineOption<std::string>("topic", 't', "Topics to record", "**")
       .defineOption<std::string>("output_bag", 'o', "output file where to write the bag");
   const auto args = std::move(desc).parse(argc, argv);
+  auto topic = args.getOption<std::string>("topic");
+  auto output_file = args.getOption<std::string>("output_bag");
 
   auto config = eolo::ipc::Config{};
-  config.topic = args.getOption<std::string>("topic");
-  auto session = eolo::ipc::zenoh::createSession(std::move(config));
+  config.topic = topic;
 
-  auto output_file = args.getOption<std::string>("output_bag");
-  auto bag_writer = eolo::bag::createMcapWriter({ .output_file = std::move(output_file) });
+  eolo::bag::ZenohRecorderParams params{
+    .session = eolo::ipc::zenoh::createSession(std::move(config)),
+    .bag_writer = eolo::bag::createMcapWriter({ .output_file = std::move(output_file) }),
+    .topics_filter_params =
+        eolo::bag::TopicFilterParams{ .include_topics_only = {}, .prefix = topic, .exclude_topics = {} }
+  };
 
-  eolo::bag::ZenohRecorderParams params{ .messages_queue_size =
-                                             eolo::bag::ZenohRecorderParams::DEFAULT_MESSAGES_QUEUE_SIZE,
-                                         .session = session,
-                                         .topics_filter_params =
-                                             eolo::bag::TopicFilterParams{ .include_topics_only = {},
-                                                                           .prefix = session->config.topic,
-                                                                           .exclude_topics = {} } };
-
-  auto zeno_recorder = eolo::bag::ZenohRecorder::create(std::move(bag_writer), std::move(params));
+  auto zeno_recorder = eolo::bag::ZenohRecorder::create(std::move(params));
   zeno_recorder.start().wait();
 
   stop_flag.wait(false);
