+++ modules/bag/src/zenoh_recorder.cpp
@@ -5,17 +5,18 @@
 #include "eolo/bag/zenoh_recorder.h"
 
 #include <future>
-#include <mutex>
+#include <memory>
 #include <utility>
 
 #include <absl/base/thread_annotations.h>
+#include <absl/synchronization/mutex.h>
 #include <fmt/core.h>
 
+#include "eolo/bag/topic_filter.h"
 #include "eolo/bag/writer.h"
-#include "eolo/containers/blocking_queue.h"
 #include "eolo/ipc/topic_database.h"
+#include "eolo/ipc/zenoh/liveliness.h"
 #include "eolo/ipc/zenoh/subscriber.h"
-#include "zenoh_recorder_subscribers_manager.h"
 
 namespace eolo::bag {
 
@@ -23,40 +24,47 @@ using BagRecord = std::pair<ipc::MessageMetadata, std::vector<std::byte>>;
 
 class ZenohRecorder::Impl {
 public:
-  Impl(std::unique_ptr<IBagWriter> writer, ZenohRecorderParams params);
+  explicit Impl(ZenohRecorderParams params);
 
   [[nodiscard]] auto start() -> std::future<void>;
 
   [[nodiscard]] auto stop() -> std::future<void>;
 
 private:
-  void createSubscriber();
-  void startRecordingBlocking();
+  void onPublisher(const ipc::zenoh::PublisherInfo& info);
+  void onPublisherAdded(const ipc::zenoh::PublisherInfo& info);
+  void onPublisherDropped(const ipc::zenoh::PublisherInfo& info);
 
 private:
-  std::unique_ptr<IBagWriter> bag_writer_;
-  ZenohRecorderParams params_;
+  std::unique_ptr<IBagWriter> bag_writer_ ABSL_GUARDED_BY(writer_mutex_);
+  absl::Mutex writer_mutex_;
 
-  containers::BlockingQueue<BagRecord> messages_;
+  TopicFilter topic_filter_;
 
+  ipc::zenoh::SessionPtr session_;
+  std::unordered_map<std::string, std::unique_ptr<ipc::zenoh::Subscriber>>
+      subscribers_ ABSL_GUARDED_BY(subscribers_mutex_);
+  absl::Mutex subscribers_mutex_;
+
+  ipc::zenoh::SessionPtr topic_info_query_session_;  // Session used to query topic service.
   std::unique_ptr<ipc::ITopicDatabase> topic_db_;
 
-  std::future<void> job_;
-
-  std::unordered_map<std::string, std::once_flag> topic_flags_ ABSL_GUARDED_BY(flags_mutex_);
-  std::mutex flags_mutex_;
+  ipc::zenoh::SessionPtr publisher_discovery_session_;
+  std::unique_ptr<ipc::zenoh::PublisherDiscovery> discover_publishers_;
 };
 
-ZenohRecorder::Impl::Impl(std::unique_ptr<IBagWriter> writer, ZenohRecorderParams params)
-  : bag_writer_(std::move(writer))
-  , params_(std::move(params))
-  , messages_(params_.messages_queue_size)
-  , topic_db_(ipc::createZenohTopicDatabase(params_.session)) {
+ZenohRecorder::Impl::Impl(ZenohRecorderParams params)
+  : bag_writer_(std::move(params.bag_writer))
+  , topic_filter_(TopicFilter::create(params.topics_filter_params))
+  , session_(std::move(params.session))
+  , topic_info_query_session_(ipc::zenoh::createSession({}))
+  , topic_db_(ipc::createZenohTopicDatabase(topic_info_query_session_))
+  , publisher_discovery_session_(ipc::zenoh::createSession({})) {
 }
 
 auto ZenohRecorder::Impl::start() -> std::future<void> {
-  createSubscriber();
-  job_ = std::async(std::launch::async, [this]() { startRecordingBlocking(); });
+  discover_publishers_ = std::make_unique<ipc::zenoh::PublisherDiscovery>(
+      publisher_discovery_session_, [this](const ipc::zenoh::PublisherInfo& info) { onPublisher(info); });
 
   std::promise<void> promise;
   promise.set_value();
@@ -64,57 +72,66 @@ auto ZenohRecorder::Impl::start() -> std::future<void> {
 }
 
 auto ZenohRecorder::Impl::stop() -> std::future<void> {
-  // Kill the subscriber.
-  // Wait for all the data to be exhausted.
-  return std::async(std::launch::async, [this] {
-    // TODO: check that killing the subscriber actually stop receiving msgs.
-    // subscriber_ = nullptr;
-    messages_.stop();
-    job_.wait();
-  });
+  discover_publishers_ = nullptr;
+  subscribers_.erase(subscribers_.begin(), subscribers_.end());
+
+  std::promise<void> promise;
+  promise.set_value();
+  return promise.get_future();
 }
 
-void ZenohRecorder::Impl::createSubscriber() {
-  // Here we leverage the fact that zenoh allows to subscribe to multiple topic via a single subscriber by
-  // using wildcards.
-  // auto cb = [this](const ipc::MessageMetadata& metadata, std::span<const std::byte> buffer) {
-  //   messages_.forceEmplace(metadata, std::vector<std::byte>{ buffer.begin(), buffer.end() });
-  // };
-
-  // TODO: use topic filter and add subscribers accordingly.
-  params_.session->config.topic = params_.topics_filter_params.prefix;
-  // subscriber_ = std::make_unique<ipc::zenoh::Subscriber>(params_.session, std::move(cb));
-}
-
-void ZenohRecorder::Impl::startRecordingBlocking() {
-  while (true) {
-    auto message = messages_.waitAndPop();
-    if (!message) {
-      break;
-    }
-
-    const auto& metadata = message->first;
-    {
-      // TODO: find a better way to do this as we don't want to block execution in this loop.
-      std::unique_lock<std::mutex> lock(flags_mutex_);  // This is need to protect topic_flags_.
-      std::call_once(topic_flags_[metadata.topic], [this, &metadata]() {
-        // TODO: replace with log.
-        fmt::println("Recording topic: {}", metadata.topic);
-        auto type_info = topic_db_->getTypeInfo(metadata.topic);
-        bag_writer_->registerSchema(type_info);
-        bag_writer_->registerChannel(metadata.topic, type_info);
-      });
-    }
-
-    fmt::println("Write msg to bag for topic: {}", metadata.topic);
-    bag_writer_->writeRecord(metadata, message->second);
+void ZenohRecorder::Impl::onPublisher(const ipc::zenoh::PublisherInfo& info) {
+  if (!topic_filter_.isAcceptable(info.topic)) {
+    return;
   }
+
+  switch (info.status) {
+    case ipc::zenoh::PublisherStatus::ALIVE:
+      onPublisherAdded(info);
+      break;
+    case ipc::zenoh::PublisherStatus::DROPPED:
+      onPublisherDropped(info);
+      break;
+  }
+}
+
+void ZenohRecorder::Impl::onPublisherAdded(const ipc::zenoh::PublisherInfo& info) {
+  auto type_info = topic_db_->getTypeInfo(info.topic);
+  {
+    absl::MutexLock lock{ &writer_mutex_ };
+    bag_writer_->registerSchema(type_info);
+    bag_writer_->registerChannel(info.topic, type_info);
+  }
+
+  auto cb = [this](const ipc::MessageMetadata& metadata, std::span<const std::byte> data) {
+    // TODO: Hate that we need a mutex here, but the discovery needs to happen in a different session,
+    // otherwise when looking for the type we may block.
+    absl::MutexLock lock{ &writer_mutex_ };
+    bag_writer_->writeRecord(metadata, data);
+  };
+  {
+    absl::MutexLock lock{ &subscribers_mutex_ };
+    throwExceptionIf<InvalidOperationException>(
+        subscribers_.contains(info.topic),
+        std::format("adding subscriber for topic: {}, but one already exists", info.topic));
+
+    subscribers_[info.topic] = std::make_unique<ipc::zenoh::Subscriber>(session_, std::move(cb));
+  }
+}
+
+void ZenohRecorder::Impl::onPublisherDropped(const ipc::zenoh::PublisherInfo& info) {
+  absl::MutexLock lock{ &subscribers_mutex_ };
+  throwExceptionIf<InvalidOperationException>(
+      !subscribers_.contains(info.topic),
+      std::format("trying to stop recording from dropped topic {}, but subscriber doesn't exist",
+                  info.topic));
+  subscribers_[info.topic] = nullptr;
+  subscribers_.extract(info.topic);
 }
 
 // ----------------------------------------------------------------------------------------------------------
 
-ZenohRecorder::ZenohRecorder(std::unique_ptr<IBagWriter> writer, ZenohRecorderParams params)
-  : impl_(std::make_unique<Impl>(std::move(writer), std::move(params))) {
+ZenohRecorder::ZenohRecorder(ZenohRecorderParams params) : impl_(std::make_unique<Impl>(std::move(params))) {
 }
 
 ZenohRecorder::~ZenohRecorder() = default;
@@ -127,8 +144,8 @@ auto ZenohRecorder::stop() -> std::future<void> {
   return impl_->stop();
 }
 
-auto ZenohRecorder::create(std::unique_ptr<IBagWriter> writer, ZenohRecorderParams params) -> ZenohRecorder {
-  return { std::move(writer), std::move(params) };
+auto ZenohRecorder::create(ZenohRecorderParams params) -> ZenohRecorder {
+  return ZenohRecorder{ std::move(params) };
 }
 
 }  // namespace eolo::bag
